chore: fix clippy lints across leaf crates and sync beads tracker
Leaf crate lint fixes:
- Remove redundant .max(0) on unsigned types (osc.rs, lineedit/mod.rs)
- Use C string literals (c"charset") instead of b"charset\0" casts
  in fontconfig FFI (fcwrap.rs)
- Qualify mem::transmute with explicit types (fcwrap.rs)
- Replace manual loop counters with zip-range idioms (luahelper,
  term/screen.rs, terminalstate/performer.rs)
- Remove redundant .into_iter() calls (harfbuzz-wasm, terminfo.rs)
- Remove unnecessary cast (umask)
- Allow clippy::too_many_arguments in toast-notification dbus

Syncs .beads/issues.jsonl with current tracker state.

+++ frankenterm/umask/src/lib.rs
@@ -737,7 +737,7 @@ mod tests {
         let original = unsafe { umask(0o055) };
         let _saver = UmaskSaver::new();
         let saved = UmaskSaver::saved_umask().unwrap();
-        assert_eq!(saved as u32, 0o055u32);
+        assert_eq!(saved, 0o055u32);
         drop(_saver);
         unsafe { umask(original) };
     }
